requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.3
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
geopandas==1.1.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
inflect==7.5.0
JayDeBeApi==1.2.3
Jinja2==3.1.6
joblib==1.5.3
jpype1==1.6.0
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
numpy==2.4.2
openai>=1.0.0
packaging==26.0
pandas==3.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyodbc==5.3.0
pyogrio==0.12.1
pyproj==3.7.2
pyspnego==0.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.2.0
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.52.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smbprotocol==1.16.0
sqlacodegen==3.2.0
SQLAlchemy==2.0.46
sqlmodel==0.0.33
starlette==0.52.1
threadpoolctl==3.6.0
typeguard==4.4.4
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
